Remove exasol.toolbox.project module
